RtMidi backend now has true blocking instead of poll and wait.
This is done by setting up a callback handler to feed a queue. The
receive() method can then do a blocking get() on this queue.

The downside is that you can't use CTRL-c to block a program that's
waiting for messages. I don't want to merge the changes until I've
found a way to get around this.