Bugfix: end_of_track messages in MIDI files were not handled correctly.
(Reported by Colin Raffel, issue #62).

The parser interpreted an end_of_track as an end of track marker in
the file. This is wrong. The end_of_track message is only used as a
timestamp to mark the time when a track ends so playback and looping
will work correctly. It is not a part of the file structure.

This caused problems when there were multiple end_of_track messages in
a track or when there were other messages following end_of_track.

The parser now treats the end_of_track messages like any other message.