Add different time units and relative/absolute timimg capability to MidiFile.
The unit (MidiFile.unit) holds a string indicating which unit to use to
display event timings. Possible values are ticks, seconds and beats.

Furthermore the timings can be viewed as absolute values or relative to
the preceeding event as indicated by MidiFile.timing.

To support timing information in beats, time signature event handling was
added to MidiFile.__iter__.