Fix HTML documentation menu
The HTML menu was missing proper sections.
This is achieved with `:caption:` in a `.. toctree`.
Since the TOC lives in the menu, added `hidden:` to not render the TOC
sections in the HTML index page.
Unfortunately this breaks the PDF, so I used a workaround with literal
latex parts which both works and is more in line with the semantics.
This unfortunately means repeating the section name in both the latex
part and the toctree caption.